moved ron method to the base_recipe
